unused implementation in build gradles, solved problems with imports and created two contr ollers instead of one, SRP
