+++ context/book/domain/build.gradle.kts
@@ -3,6 +3,5 @@ plugins {
 }
 
 dependencies {
-    implementation("org.springframework.boot:spring-boot-starter-web:2.3.3.RELEASE")
 }
