+++ context/book/primary-adapter/rest/spring-web-mvc/src/main/kotlin/si/dime/kotlin/tutorials/rest/booklibrary/book/primaryadapter/rest/BookDocument.kt
@@ -1,0 +1,5 @@
+package si.dime.kotlin.tutorials.rest.booklibrary.book.primaryadapter.rest
+
+import si.dime.kotlin.tutorials.rest.booklibrary.book.domain.Book
+
+data class BookDocument (val data: List<Book>)
