+++ context/book/secondary-adapter/build.gradle.kts
@@ -3,6 +3,6 @@ plugins {
 }
 
 dependencies {
-    implementation("org.springframework.boot:spring-boot-starter-web:2.3.3.RELEASE")
     implementation(project(":context:book:domain"))
+    implementation("org.springframework.boot:spring-boot-starter-web:2.3.3.RELEASE")
 }
