+++ library-source/command-query/src/main/kotlin/si/dime/kotlin/tutorials/rest/booklibrary/librarysource/commandquery/query/QueryBus.kt
@@ -1,6 +1,7 @@
 package si.dime.kotlin.tutorials.rest.booklibrary.librarysource.commandquery.query
 
 import arrow.core.Either
+import si.dime.kotlin.tutorials.rest.booklibrary.librarysource.fp.catch
 
 import kotlin.reflect.KClass
