+++ library-source/fp/build.gradle.kts
@@ -4,6 +4,4 @@ plugins {
 
 dependencies {
     api("io.arrow-kt:arrow-core:0.11.0")
-    api("io.arrow-kt:arrow-fx:0.11.0")
-    api("io.arrow-kt:arrow-syntax:0.11.0")
 }
